+++ .idea/deploymentTargetDropDown.xml
@@ -1,19 +1,17 @@
 <?xml version="1.0" encoding="UTF-8"?>
 <project version="4">
   <component name="deploymentTargetDropDown">
-    <runningDeviceTargetSelectedWithDropDown>
+    <targetSelectedWithDropDown>
       <Target>
-        <type value="RUNNING_DEVICE_TARGET" />
+        <type value="QUICK_BOOT_TARGET" />
         <deviceKey>
           <Key>
             <type value="VIRTUAL_DEVICE_PATH" />
-            <value value="$PROJECT_DIR$/../.android/avd/Pixel_4_API_33.avd" />
+            <value value="$USER_HOME$/.android/avd/Nexus_7_2012_API_32.avd" />
           </Key>
         </deviceKey>
       </Target>
-    </runningDeviceTargetSelectedWithDropDown>
-    <timeTargetWasSelectedWithDropDown value="2022-11-10T02:55:53.508511Z" />
+    </targetSelectedWithDropDown>
+    <timeTargetWasSelectedWithDropDown value="2022-11-10T03:27:36.477241Z" />
   </component>
-    <timeTargetWasSelectedWithDropDown value="2022-11-10T02:26:43.595836Z" />
-    <timeTargetWasSelectedWithDropDown value="2022-11-10T01:16:55.899893Z" />
 </project>
