+++ app/src/androidTest/java/com/emich/vibe/ExampleInstrumentedTest.java
@@ -1,4 +1,4 @@
-package com.example.vibe;
+package com.emich.vibe;
 
 import android.content.Context;
