+++ app/src/main/AndroidManifest.xml
@@ -15,18 +15,18 @@
         android:theme="@style/Theme.LoginPage"
         tools:targetApi="31">
         <activity
-            android:name="com.example.vibe.ChatLog"
+            android:name="com.emich.vibe.ChatLog"
             android:exported="false"
             android:theme="@style/Theme.Design.NoActionBar"/>
         <activity
-            android:name="com.example.vibe.Settings_View"
+            android:name="com.emich.vibe.Settings_View"
             android:exported="false" />
         <activity
-            android:name="com.example.vibe.ConversationView"
+            android:name="com.emich.vibe.ConversationView"
             android:exported="false"
             android:theme="@style/Theme.Design.NoActionBar"/>
         <activity
-            android:name="com.example.vibe.SplashScreen"
+            android:name="com.emich.vibe.SplashScreen"
             android:exported="true">
             <intent-filter>
                 <action android:name="android.intent.action.MAIN" />
@@ -35,22 +35,22 @@
             </intent-filter>
         </activity>
         <activity
-            android:name="com.example.vibe.Register"
+            android:name="com.emich.vibe.Register"
             android:exported="false" />
         <activity
-            android:name="com.example.vibe.FriendSearch"
+            android:name="com.emich.vibe.FriendSearch"
             android:exported="false"
             android:theme="@style/Theme.Design.NoActionBar"/>
 
         <activity
-            android:name="com.example.vibe.ChangePassword"
+            android:name="com.emich.vibe.ChangePassword"
             android:exported="false">
         <meta-data
             android:name="android.app.lib_name"
             android:value="" />
         </activity>
         <activity
-            android:name="com.example.vibe.Login"
+            android:name="com.emich.vibe.Login"
             android:exported="true">
 <!--            <intent-filter>-->
 <!--                <action android:name="android.intent.action.MAIN" />-->
@@ -59,7 +59,7 @@
 <!--            </intent-filter>-->
         </activity>
         <activity
-            android:name="com.example.vibe.MainActivity"
+            android:name="com.emich.vibe.MainActivity"
             android:exported="false">
         </activity>
     </application>
