+++ app/src/main/java/com/emich/vibe/ChatsProvider.java
@@ -1,27 +1,20 @@
-package com.example.vibe;
+package com.emich.vibe;
 
 import com.google.android.gms.tasks.Task;
 import com.google.firebase.firestore.CollectionReference;
-import com.google.firebase.firestore.DocumentReference;
 import com.google.firebase.firestore.FirebaseFirestore;
 import com.google.firebase.firestore.Query;
 
 import java.util.ArrayList;
 
-
-public class MessagesProvider {
+public class ChatsProvider {
 
     CollectionReference mCollection;
 
-    public MessagesProvider() {
-        mCollection = FirebaseFirestore.getInstance().collection("messages");
+    public ChatsProvider() {
+        mCollection = FirebaseFirestore.getInstance().collection("chats");
     }
 
-    public Task<Void> create(Message message) {
-        DocumentReference document = mCollection.document();
-        message.setId(document.getId());
-        return document.set(message);
-    }
 
     public Query getUserChats(String idUser) {
         return mCollection.whereArrayContains("ids", idUser);
