+++ app/src/main/java/com/emich/vibe/FriendSearch.java
@@ -1,4 +1,4 @@
-package com.example.vibe;
+package com.emich.vibe;
 
 import android.content.Intent;
 import android.os.Bundle;
@@ -7,7 +7,7 @@ import android.view.MenuItem;
 import androidx.appcompat.app.AppCompatActivity;
 import androidx.appcompat.widget.Toolbar;
 
-import com.example.vibe.R;
+import com.emich.vibe.R;
 
 public class FriendSearch extends AppCompatActivity {
     @Override
