+++ app/src/main/java/com/emich/vibe/Message.java
@@ -1,4 +1,4 @@
-package com.example.vibe;
+package com.emich.vibe;
 
 import java.text.SimpleDateFormat;
 import java.util.Date;
