+++ app/src/main/java/com/emich/vibe/MessagesProvider.java
@@ -1,20 +1,27 @@
-package com.example.vibe;
+package com.emich.vibe;
 
 import com.google.android.gms.tasks.Task;
 import com.google.firebase.firestore.CollectionReference;
+import com.google.firebase.firestore.DocumentReference;
 import com.google.firebase.firestore.FirebaseFirestore;
 import com.google.firebase.firestore.Query;
 
 import java.util.ArrayList;
 
-public class ChatsProvider {
+
+public class MessagesProvider {
 
     CollectionReference mCollection;
 
-    public ChatsProvider() {
-        mCollection = FirebaseFirestore.getInstance().collection("chats");
+    public MessagesProvider() {
+        mCollection = FirebaseFirestore.getInstance().collection("messages");
     }
 
+    public Task<Void> create(Message message) {
+        DocumentReference document = mCollection.document();
+        message.setId(document.getId());
+        return document.set(message);
+    }
 
     public Query getUserChats(String idUser) {
         return mCollection.whereArrayContains("ids", idUser);
