+++ app/src/main/java/com/emich/vibe/UserAdapter.java
@@ -1,4 +1,4 @@
-package com.example.vibe;
+package com.emich.vibe;
 
 import android.annotation.SuppressLint;
 import android.content.Context;
@@ -11,6 +11,7 @@ import android.widget.TextView;
 import androidx.annotation.NonNull;
 import androidx.recyclerview.widget.RecyclerView;
 
+import com.emich.vibe.R;
 import com.google.firebase.firestore.auth.User;
 
 import java.util.List;
