+++ app/src/main/java/com/emich/vibe/messages/User.java
@@ -1,4 +1,4 @@
-package com.example.vibe.messages;
+package com.emich.vibe.messages;
 
 public class User {
