+++ app/src/main/res/layout/activity_change_password.xml
@@ -7,7 +7,7 @@
     android:layout_height="match_parent"
     android:layout_gravity="center"
     android:fitsSystemWindows="true"
-    tools:context=".ChangePassword">
+    tools:context="com.emich.vibe.ChangePassword">
 
     <TextView
         android:id="@+id/textView"
