+++ app/src/main/res/layout/activity_chat_log.xml
@@ -5,7 +5,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:background="@drawable/backgroundlight"
-    tools:context=".ChatLog">
+    tools:context="com.emich.vibe.ChatLog">
 
 
     <androidx.appcompat.widget.Toolbar
