+++ app/src/main/res/layout/activity_register.xml
@@ -5,7 +5,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:background="@drawable/background"
-    tools:context=".Register">
+    tools:context="com.emich.vibe.Register">
 
     <TextView
         android:id="@+id/TextView"
