+++ app/src/main/res/layout/activity_settings_view.xml
@@ -4,7 +4,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".Settings_View">
+    tools:context="com.emich.vibe.Settings_View">
 
 
     <Button
