+++ app/src/main/res/layout/activity_splash_screen.xml
@@ -5,7 +5,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:background="@drawable/background"
-    tools:context=".SplashScreen">
+    tools:context="com.emich.vibe.SplashScreen">
 
     <ImageView
         android:id="@+id/image"
